修复两个直播间小问题 (#629)
礼物id的int越界修复
弹幕延迟计算防止除以0

+++ src/BiliLite.UWP/Models/Common/Live/LiveGiftItem.cs
@@ -128,7 +128,7 @@ namespace BiliLite.Models.Common.Live
         public int ComboResourcesId { get; set; }
 
         [JsonProperty("goods_id")]
-        public int GoodsId { get; set; }
+        public long GoodsId { get; set; }
 
         public int Num { get; set; } = 1;
     }
